+++ fabric/src/main/java/io/github/sakurawald/fuji/module/initializer/echo/send_dialog/SendDialogInitializer.java
@@ -1,0 +1,75 @@
+package io.github.sakurawald.fuji.module.initializer.echo.send_dialog;
+
+import io.github.sakurawald.fuji.core.auxiliary.minecraft.CommandHelper;
+import io.github.sakurawald.fuji.core.auxiliary.minecraft.TextHelper;
+import io.github.sakurawald.fuji.core.command.annotation.CommandNode;
+import io.github.sakurawald.fuji.core.command.annotation.CommandRequirement;
+import io.github.sakurawald.fuji.core.command.annotation.CommandSource;
+import io.github.sakurawald.fuji.core.command.argument.wrapper.impl.GreedyString;
+import io.github.sakurawald.fuji.core.document.annotation.Document;
+import io.github.sakurawald.fuji.module.initializer.ModuleInitializer;
+import io.github.sakurawald.fuji.module.initializer.echo.send_dialog.structure.DialogGui;
+import java.util.Optional;
+import net.minecraft.item.Item;
+import net.minecraft.item.Items;
+import net.minecraft.server.command.ServerCommandSource;
+import net.minecraft.server.network.ServerPlayerEntity;
+import net.minecraft.text.Text;
+
+@Document("""
+    Send text using the `dialog GUI`.
+
+    For example:
+    1. `/send-dialog Steve \\<blue\\>Hello`
+    2. `/send-dialog Steve --noButtonSlotIndex -1 \\<green\\>Confirm me.`
+    3. `/send-dialog Steve --yesButtonCommand "say confirmed" \\<blue\\>Confirm me.`
+    """)
+
+public class SendDialogInitializer extends ModuleInitializer {
+
+    @Document("Send a dialog GUI to a player.")
+    @CommandNode("send-dialog")
+    @CommandRequirement(level = 4)
+    private static int $sendDialog(@CommandSource ServerCommandSource source
+        , ServerPlayerEntity player
+        , Optional<Integer> rows
+        , Optional<Integer> yesButtonSlotIndex
+        , Optional<Item> yesButtonItem
+        , Optional<String> yesButtonName
+        , Optional<String> yesButtonCommand
+        , Optional<Integer> noButtonSlotIndex
+        , Optional<Item> noButtonItem
+        , Optional<String> noButtonName
+        , Optional<String> noButtonCommand
+        , Optional<Boolean> canCloseUsingNoButton
+        , GreedyString title) {
+
+        /* Extract the arguments. */
+        Integer $rows = rows.orElse(1);
+        String $title = title.getValue();
+        Text dialogTitleText = TextHelper.getTextByValue(player, $title);
+
+        int $yesButtonSlotIndex = yesButtonSlotIndex.orElse(2);
+        Item $yesButtonItem = yesButtonItem.orElse(Items.SLIME_BALL);
+        String $yesButtonName = yesButtonName.orElse("<green><b>YES");
+        Text $$yesButtonName = TextHelper.getTextByValue(player, $yesButtonName);
+        String $yesButtonCommand = yesButtonCommand.orElse("");
+
+        int $noButtonSlotIndex = noButtonSlotIndex.orElse(6);
+        Item $noButtonItem = noButtonItem.orElse(Items.MAGMA_CREAM);
+        String $noButtonName = noButtonName.orElse("<dark_red><b>NO");
+        Text $$noButtonName = TextHelper.getTextByValue(player, $noButtonName);
+        String $noButtonCommand = noButtonCommand.orElse("");
+        boolean $canCloseUsingNoButton = canCloseUsingNoButton.orElse(false);
+
+        /* Make the dialog GUI. */
+        DialogGui.makeDialogGui($rows, player, dialogTitleText
+                , $yesButtonSlotIndex, $yesButtonItem, $$yesButtonName, $yesButtonCommand
+                , $noButtonSlotIndex, $noButtonItem, $$noButtonName, $noButtonCommand, $canCloseUsingNoButton
+            )
+            .open();
+
+        return CommandHelper.Return.SUCCESS;
+    }
+
+}